requirements.txt
Django==1.6.2
Werkzeug==0.9.4
argparse==1.2.1
distribute==0.7.3
dj-database-url==0.3.0
dj-static==0.0.5
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-toolbelt==0.0.1
gunicorn==18.0
icalendar==3.6.1
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
pytz==2014.2
requests==2.2.1
six==1.5.2
sqlparse==0.1.11
static==1.0.2
uWSGI==2.0.2
wsgiref==0.1.2
